pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "print3d-skill"
version = "0.1.0"
description = "AI agent skill for full-stack 3D printing assistance"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "trimesh>=4.0",
    "manifold3d>=3.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "Pillow>=10.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
openscad = []
slicer = []
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.5",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"print3d_skill.knowledge_base" = ["*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
